added streamer class
